bot/requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
openai==1.51.0
httpx==0.27.2
python-dotenv==1.0.1
